requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
numpy==2.1.0
Pillow==10.4.0
Requests==2.32.3
tensorflow==2.17.0
tensorflow_intel==2.17.0
pickle
